♻️ Refator: reducer 및 유효성 검사 로직 커스텀 훅으로 분리
